Готовый проект(без шлифовки)

+++ Models/Counters/Tariff.cs
@@ -9,8 +9,8 @@ namespace ERCTest.Models.Counters
         public int Id { get; set; }
 
         public string ServiceName { get; set; }
-        public double TariffPrice { get; set; }
-        public double TariffWithoutCouner { get; set; }
+        public decimal TariffPrice { get; set; }
+        public decimal TariffWithoutCouner { get; set; }
         public string UnitOfMeasurment { get; set; }
     }
 }
